Refactor favorite vessels component
