+++ frontend/src/App.tsx
@@ -12,7 +12,7 @@ import { APIWorker } from './api/APIWorker'
 import { BackofficeMode } from './api/BackofficeMode'
 import NamespaceContext from './domain/context/NamespaceContext'
 import { ErrorToastNotification } from './features/commonComponents/ErrorToastNotification'
-import FavoriteVessels from './features/favorite_vessels/FavoriteVessels'
+import { FavoriteVessels } from './features/FavoriteVessels'
 import { Healthcheck } from './features/Healthcheck'
 import LayersSidebar from './features/layers/LayersSidebar'
 import { DrawLayerModal } from './features/map/draw/DrawModal'
